moved access_filters functions to separate filter_helpers.py script, added documentation, and comparison tested with preexisting full voyager run on juno
